Changes to stock: added a variable to store investment value. also changed invested balance variable name to more clearly differentiate them. changed the getters and setter to match
changes to GUI: added a loading screen because the database and simulator takes some time to load in. havent got it to work completely so theres still a delay.

I also coded up most of the stock display window. need to add actual trading features. also need to fix a bug because the stock buttons dont open the stock window and im not sure why